- Remove UI system (too hard to use and UI should be specific to a game not to an engine)
- Added ImGui for debug UI (much easier to use)
- Added vkCmdSetViewport
- Added vkCmdSetScissor
- Added vkCmdPushConstants
- Added ability to create texture from RGBA data in AssetFetcher
- Added instanced Vulkan procedures (used by ImGui, should not be used for non internal use)

+++ code/src/Graphics/Vulkan/QueryPool.cpp
@@ -3,14 +3,14 @@
 Pu::QueryPool::QueryPool(LogicalDevice & device, QueryType type, size_t count)
 	: parent(&device)
 {
-	const QueryPoolCreateInfo createInfo(type, count);
+	const QueryPoolCreateInfo createInfo(type, static_cast<uint32>(count));
 	VK_VALIDATE(parent->vkCreateQueryPool(parent->hndl, &createInfo, nullptr, &hndl), PFN_vkCreateQueryPool);
 }
 
 Pu::QueryPool::QueryPool(LogicalDevice & device, size_t count, QueryPipelineStatisticFlag statistics)
 	: parent(&device)
 {
-	const QueryPoolCreateInfo createInfo(QueryType::PipelineStatistics, count, statistics);
+	const QueryPoolCreateInfo createInfo(QueryType::PipelineStatistics, static_cast<uint32>(count), statistics);
 	VK_VALIDATE(parent->vkCreateQueryPool(parent->hndl, &createInfo, nullptr, &hndl), PFN_vkCreateQueryPool);
 }
